Organized the root folder a little more
